ml_project/requirements.txt
black==20.8b1
hydra-core==1.0.6
jupyterlab==3.0.13
pandas==1.2.3
pandas-profiling==2.11.0
pytest==6.2.3
scikit-learn==0.24.1